+++ packages/components-angular/src/app/examples/index.ts
@@ -21,6 +21,8 @@ import { DrilldownExampleComponent } from './drilldown-example.component';
 import { HiddenAttributeExampleComponent } from './hidden-attribute-example.component';
 import { InlineNotificationExampleActionButtonComponent } from './inline-notification-example-action-button.component';
 import { InlineNotificationExampleEventsComponent } from './inline-notification-example-events.component';
+import { InputPasswordExampleControlledComponent } from './input-password-example-controlled.component';
+import { InputPasswordExampleComponent } from './input-password-example.component';
 import { LinkTileProductExampleComponent } from './link-tile-product-example.component';
 import { ModalExampleComponent } from './modal-example.component';
 import { MultiSelectExampleControlledComponent } from './multi-select-example-controlled.component';
@@ -75,6 +77,8 @@ export const examples = [
   HiddenAttributeExampleComponent,
   InlineNotificationExampleActionButtonComponent,
   InlineNotificationExampleEventsComponent,
+  InputPasswordExampleControlledComponent,
+  InputPasswordExampleComponent,
   LinkTileProductExampleComponent,
   ModalExampleComponent,
   MultiSelectExampleComponent,
@@ -128,6 +132,8 @@ export * from './drilldown-example-custom-content.component';
 export * from './hidden-attribute-example.component';
 export * from './inline-notification-example-action-button.component';
 export * from './inline-notification-example-events.component';
+export * from './input-password-example.component';
+export * from './input-password-example-controlled.component';
 export * from './link-tile-product-example.component';
 export * from './modal-example.component';
 export * from './multi-select-example.component';